Refactor import statements for Referencer to improve module structure
